Move exception classes from domain_langauge.py to semparse/common/errors.py
