+++ allennlp/data/dataset_readers/semantic_parsing/wikitables/wikitables.py
@@ -23,7 +23,7 @@ from allennlp.data.dataset_readers.semantic_parsing.wikitables import util
 from allennlp.semparse.contexts import TableQuestionKnowledgeGraph
 from allennlp.semparse.type_declarations import wikitables_lambda_dcs as wt_types
 from allennlp.semparse.worlds import WikiTablesWorld
-from allennlp.semparse.worlds.world import ParsingError
+from allennlp.semparse.common.errors import ParsingError
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
