+++ allennlp/semparse/domain_languages/__init__.py
@@ -1,6 +1,6 @@
-from allennlp.semparse.domain_languages.domain_language import (DomainLanguage, ParsingError,
-                                                                ExecutionError, START_SYMBOL,
+from allennlp.semparse.domain_languages.domain_language import (DomainLanguage, START_SYMBOL,
                                                                 predicate, predicate_with_side_args)
+from allennlp.semparse.common.errors import ParsingError, ExecutionError
 from allennlp.semparse.domain_languages.nlvr_language import NlvrLanguage
 from allennlp.semparse.domain_languages.quarel_language import QuaRelLanguage
 from allennlp.semparse.domain_languages.wikitables_language import WikiTablesLanguage
