+++ allennlp/semparse/domain_languages/domain_language.py
@@ -10,6 +10,7 @@ from nltk import Tree
 
 from allennlp.common.util import START_SYMBOL
 from allennlp.semparse import util
+from allennlp.semparse.common.errors import ParsingError, ExecutionError
 
 logger = logging.getLogger(__name__)
 
@@ -138,35 +139,6 @@ class FunctionType(PredicateType):
         if isinstance(self, other.__class__):
             return self.name == other.name
         return NotImplemented
-
-
-class ParsingError(Exception):
-    """
-    This exception gets raised when there is a parsing error during logical form processing.  This
-    might happen because you're not handling the full set of possible logical forms, for instance,
-    and having this error provides a consistent way to catch those errors and log how frequently
-    this occurs.
-    """
-    def __init__(self, message):
-        super().__init__()
-        self.message = message
-
-    def __str__(self):
-        return repr(self.message)
-
-
-class ExecutionError(Exception):
-    """
-    This exception gets raised when you're trying to execute a logical form that your executor does
-    not understand. This may be because your logical form contains a function with an invalid name
-    or a set of arguments whose types do not match those that the function expects.
-    """
-    def __init__(self, message):
-        super().__init__()
-        self.message = message
-
-    def __str__(self):
-        return repr(self.message)
 
 
 def predicate(function: Callable) -> Callable:  # pylint: disable=invalid-name
