+++ allennlp/semparse/worlds/world.py
@@ -7,9 +7,9 @@ from nltk import Tree
 from nltk.sem.logic import ApplicationExpression, Expression, LambdaExpression, BasicType, Type
 
 from allennlp.semparse.type_declarations import type_declaration as types
-from allennlp.semparse.domain_languages.domain_language import ParsingError
 from allennlp.semparse.domain_languages.domain_language import nltk_tree_to_logical_form
 from allennlp.semparse import util as semparse_util
+from allennlp.semparse.common.errors import ParsingError
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
