+++ allennlp/tests/semparse/worlds/wikitables_world_test.py
@@ -5,7 +5,7 @@ import pytest
 
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.tokenizers import Token
-from allennlp.semparse import ParsingError
+from allennlp.semparse.common.errors import ParsingError
 from allennlp.semparse.contexts import TableQuestionKnowledgeGraph
 from allennlp.semparse.worlds import WikiTablesWorld
 from allennlp.semparse.type_declarations import wikitables_lambda_dcs as types
