+++ allennlp/tests/semparse/worlds/world_test.py
@@ -3,7 +3,8 @@ from overrides import overrides
 
 from allennlp.common.testing import AllenNlpTestCase
 from allennlp.data.tokenizers import Token
-from allennlp.semparse import ParsingError, World
+from allennlp.semparse.common.errors import ParsingError
+from allennlp.semparse import World
 from allennlp.semparse.contexts import TableQuestionKnowledgeGraph
 from allennlp.semparse.worlds import WikiTablesWorld
